KE-9016 convert higher precision value to lower precision

+++ core/src/main/java/org/apache/calcite/runtime/SqlFunctions.java
@@ -1628,7 +1628,7 @@ public class SqlFunctions {
   }
 
   public static short toShort(String s) {
-    return Short.parseShort(s.trim());
+    return (short) Double.parseDouble(s.trim());
   }
 
   public static short toShort(Number number) {
@@ -1681,7 +1681,7 @@ public class SqlFunctions {
   }
 
   public static int toInt(String s) {
-    return Integer.parseInt(s.trim());
+    return (int) Double.parseDouble(s.trim());
   }
 
   public static int toInt(Number number) {
